Show the unloaded regions in the breakpoint panel
